Move select a profile warnings to welcome component
