+++ src/app/App.blocks.test.tsx
@@ -7,8 +7,6 @@ import { AppRouter, Main } from "./App.blocks";
 import { getDefaultProfileId, render, screen, waitFor, within, env, defaultNetMocks } from "@/utils/testing-library";
 import { toasts } from "@/app/services";
 import * as useProfileSynchronizerHook from "@/app/hooks/use-profile-synchronizer";
-import { translations } from "@/app/i18n/common/i18n";
-import { ProfilePaths } from "@/router/paths";
 const history = createHashHistory();
 
 jest.setTimeout(7000);
@@ -239,64 +237,5 @@ describe("App Main", () => {
 		onProfileUpdated();
 
 		await waitFor(() => expect(history.location.pathname).toBe("/"));
-	});
-
-	describe("useDeeplink", () => {
-		const mainnetDeepLink =
-			"/?method=transfer&coin=ark&network=ark.mainnet&recipient=DNjuJEDQkhrJ7cA9FZ2iVXt5anYiM8Jtc9&amount=1.2&memo=ARK";
-		let toastWarningSpy: jest.SpyInstance;
-
-		beforeEach(() => {
-			toastWarningSpy = jest.spyOn(toasts, "warning").mockImplementation();
-		});
-
-		afterEach(() => {
-			toastWarningSpy.mockRestore();
-		});
-
-		it("should prompt the user to select a profile", async () => {
-			render(
-				<Route path="/">
-					<Main />
-				</Route>,
-				{
-					history,
-					route: mainnetDeepLink,
-					withProviders: true,
-				},
-			);
-
-			await waitFor(() =>
-				expect(toastWarningSpy).toHaveBeenCalledWith(translations.SELECT_A_PROFILE, { delay: 500 }),
-			);
-		});
-
-		it.each([
-			["createProfile", ProfilePaths.CreateProfile],
-			["importProfile", ProfilePaths.ImportProfile],
-		])("should clear deeplink and do not show a warning toast in %s page", async (page, path) => {
-			const toastWarningSpy = jest.spyOn(toasts, "warning").mockImplementation();
-
-			render(
-				<Route path="/">
-					<Main />
-				</Route>,
-				{
-					history,
-					route: mainnetDeepLink,
-					withProviders: true,
-				},
-			);
-
-			await waitFor(() => {
-				expect(toastWarningSpy).toHaveBeenCalledWith(translations.SELECT_A_PROFILE, { delay: 500 });
-			});
-
-			history.push(path);
-
-			await waitFor(() => expect(toastWarningSpy).toHaveBeenCalledTimes(1));
-
-			toastWarningSpy.mockRestore();
-		});
 	});
 });
